# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MYiTH Core API
  description: |
    채용공고와 국가직무능력표준(NCS)을 결합한 개인 맞춤형 취업 로드맵 서비스.

    인증: Google ID Token으로 로그인 후 발급받은 accessToken을
    Authorization: Bearer {token} 헤더에 실어 보낸다.

    공통 응답: 모든 성공 응답은 data로 감싼다. (GET /api/health 제외)
    공통 오류: { "error": { code, message, fieldErrors, requestId } }
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: 로컬
- url: https://api.myith.example
  description: 운영
security:
- bearerAuth: []
...
paths:
...
  /api/auth/refresh:
    post:
      tags:
      - Auth
      summary: 토큰 갱신
      description: |-
        401 + TOKEN_EXPIRED 수신 시 프론트가 이 엔드포인트를 1회 호출하고 원 요청을 재시도한다.
        재차 401이면 로그인 화면으로 보낸다.
        refreshToken은 요청 바디로 받는다(httpOnly 쿠키 아님).
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: 갱신 성공
          content:
            application/json:
              example:
                data:
                  accessToken: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c3JfMDFKM0FCQyJ9.new
        "401":
          description: INVALID_REFRESH_TOKEN — refreshToken 무효
          content:
            application/json:
              example:
                error:
                  code: INVALID_REFRESH_TOKEN
                  message: 유효하지 않은 토큰입니다.
                  requestId: req_01J3ABC
      security: []
  /api/auth/google:
    post:
      tags:
      - Auth
      summary: 구글 로그인
      description: |-
        Google ID Token을 서버가 로컬 검증한다(캐싱된 공개키로 JWT 서명 검증).
        최초 로그인 시 자동 가입하고 users.nickname을 구글 이름으로 세팅한다.
        isNewUser로 프론트가 온보딩(알 선택 화면 1-1)으로 분기한다.
      operationId: googleLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginRequest"
        required: true
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              example:
                data:
                  accessToken: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c3JfMDFKM0FCQyJ9.abc
                  refreshToken: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c3JfMDFKM0FCQyJ9.xyz
                  isNewUser: true
        "401":
          description: INVALID_ID_TOKEN — Google ID Token 검증 실패
          content:
            application/json:
              example:
                error:
                  code: INVALID_ID_TOKEN
                  message: Google ID Token 검증에 실패했습니다.
                  requestId: req_01J3ABC
      security: []
...
  /api/ai-enhancements/{requestId}:
    get:
      tags:
      - Quest
      summary: AI 보완 결과 조회
      description: |-
        POST /api/quests/{questId}/ai-enhancements에서 202 수신 후 프론트가 짧은 간격으로 폴링한다.
        enhancedStar → 비교 모달 오른쪽 / feedback → 항목별 힌트 / resumeDraft → 자기소개서 초안 영역.
        서버는 원문을 수정하지 않는다. 사용자가 '적용'을 누르면 프론트가 textarea를 채우고 PUT /star로 저장한다.
      operationId: getResult
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 실패
          content:
            application/json:
              examples:
                FAILED:
                  description: FAILED
                  value:
                    data:
                      requestId: aie_01J3ABC
                      status: FAILED
                      errorCode: AI_PROVIDER_TIMEOUT
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: 리프레시 토큰
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c3JfMDFKM0FCQyJ9.xyz
      required:
      - refreshToken
    GoogleLoginRequest:
      type: object
      properties:
        idToken:
          type: string
          description: Google에서 받은 ID Token
          example: eyJhbGciOiJSUzI1NiIs...
      required:
      - idToken
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT